requirements.txt
fastapi
httpx
loguru

openai
duckduckgo-search
langchain-community==0.3.14
langchain-chroma==0.2.0
langchain-huggingface==0.1.2
langchain-openai==0.3.1

python-dotenv
